feat: 상품 목록 조회 API 추가
- ProductService : 상품 서비스
- DiscountService : 할인 서비스
- ProductListService : 상품, 할인을 통해 상품 목록을 제공하는 서비스

+++ backend/src/main/java/com/team28/sidedish/controller/dto/ProductDetailResponse.java
@@ -30,10 +30,10 @@ public class ProductDetailResponse {
     private List<ProductImageResponse> detailImages;
 
     @Schema(description = "상품 가격")
-    private int price;
+    private Long price;
 
     @Schema(description = "할인된 가격")
-    private int discountPrice;
+    private Long discountPrice;
 
     @Schema(description = "상품 재고")
     private int stockQuantity;
